asim: move metrics_*.go into metrics pkg
This patch moves `metrics_tracker.go` and `metrics.go` into a new pkg
`metrics`. This is done to decouple the metrics code from the `asim`
pkg.

Release note: None
