+++ pkg/kv/kvserver/asim/asim_test.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim/config"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim/metrics"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim/state"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim/workload"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
@@ -36,7 +37,7 @@ func TestRunAllocatorSimulator(t *testing.T) {
 	interval := 10 * time.Second
 	rwg := make([]workload.Generator, 1)
 	rwg[0] = workload.TestCreateWorkloadGenerator(settings.Seed, start, 1, 10)
-	m := asim.NewMetricsTracker(os.Stdout)
+	m := metrics.NewMetricsTracker(os.Stdout)
 	changer := state.NewReplicaChanger()
 	s := state.LoadConfig(state.ComplexConfig)
 
@@ -79,7 +80,7 @@ func TestAllocatorSimulatorSpeed(t *testing.T) {
 		rwg := make([]workload.Generator, 1)
 		rwg[0] = workload.TestCreateWorkloadGenerator(settings.Seed, start, stores, int64(keyspace))
 		changer := state.NewReplicaChanger()
-		m := asim.NewMetricsTracker() // no output
+		m := metrics.NewMetricsTracker() // no output
 		replicaDistribution := make([]float64, stores)
 
 		// NB: Here create half of the stores with equal replica counts, the
@@ -151,7 +152,7 @@ func TestAllocatorSimulatorDeterministic(t *testing.T) {
 		rwg := make([]workload.Generator, 1)
 		rwg[0] = workload.TestCreateWorkloadGenerator(settings.Seed, start, stores, int64(keyspace))
 		changer := state.NewReplicaChanger()
-		m := asim.NewMetricsTracker() // no output
+		m := metrics.NewMetricsTracker() // no output
 		replicaDistribution := make([]float64, stores)
 
 		// NB: Here create half of the stores with equal replica counts, the
