+++ pkg/kv/kvserver/asim/metrics/BUILD.bazel
@@ -1,0 +1,29 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "metrics",
+    srcs = ["metrics_tracker.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim/metrics",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/kv/kvserver/asim/state",
+        "//pkg/util/encoding/csv",
+    ],
+)
+
+go_test(
+    name = "metrics_test",
+    srcs = ["metrics_tracker_test.go"],
+    args = ["-test.timeout=295s"],
+    deps = [
+        ":metrics",
+        "//pkg/kv/kvserver/asim",
+        "//pkg/kv/kvserver/asim/config",
+        "//pkg/kv/kvserver/asim/state",
+        "//pkg/kv/kvserver/asim/workload",
+        "@com_github_stretchr_testify//require",
+    ],
+)
+
+get_x_data(name = "get_x_data")
